src/gateway/requirements.txt
auto_mix_prep==0.2.0
fastapi==0.101.1
pydantic==1.10.10
python-dotenv==1.0.0
SQLAlchemy==2.0.19
